fix(quic)!: default to platform TLS verifier; gate skip-verify behind opt-in feature (ADR-0011)
Implements ADR-0011. Prior versions of midstreamer-quic shipped a TLS
client that unconditionally installed a `SkipServerVerification` cert
verifier in `QuicConnection::connect` (see `native.rs:39-43` before this
patch). Any consumer of the published crate inherited an MITM-vulnerable
client by default — this was a real, live security issue in 0.1.0.

This patch:

  * Switches the default verifier to `rustls-platform-verifier`, which
    validates server certificates against the OS trust store (matches
    Mac Keychain, Windows cert store, system `ca-certificates` on
    Linux). No more silent acceptance of arbitrary certs.

  * Moves the skip-verify code into a new `insecure` module that is
    `#[cfg]`-gated on the new
    `insecure-dev-only-skip-server-verification` cargo feature. The
    module — and the `SkipServerVerification` struct — does not exist
    in default builds. Enabling the feature is opt-in at the
    `[dependencies]` level and surfaces in `cargo tree --features`.

  * Emits a `tracing::warn!` on every `connect()` when the insecure
    feature is enabled, so anyone running with skip-verify by accident
    sees it in logs.

  * Drops the dead `rustls = "0.22"` direct dep that was orphaned at
    the local manifest level (quinn 0.11 already uses 0.23 internally;
    the 0.22 entry contributed to the dual-rustls duplicate-version
    finding tracked by ADR-0014).

Version bump: 0.1.0 -> 0.1.1 (security patch). The Rust semver convention
for 0.y.z treats this as a patch; the *behavior* is breaking for any
consumer who silently relied on the skip-verify default. That reliance
was itself the bug; clear callout in this commit and in the PR
description.

Build verification:
  * `cargo check -p midstreamer-quic`                             - clean
  * `cargo check -p midstreamer-quic --features insecure-...`     - clean
  * `cargo check -p midstreamer-quic --tests`                     - clean
  * `cargo check -p midstreamer-quic --tests --features insecure` - clean
